Reorganize dashboard component and sass files
