+++ src/sass/landing 2.scss
@@ -2,7 +2,7 @@
 
 
 .landing-body{
-    background-image: url('../imgs/landingHero.jpg');
+    background-image: url('../assets/imgs/landingHero.jpg');
     background-size:cover;
     background-repeat: no-repeat;
     font-family: "Montserrat", sans-serif;
