+++ src/sass/landing.scss
@@ -1,7 +1,7 @@
 @import url('https://fonts.googleapis.com/css?family=Montserrat&display=swap');
 
 .landing-body {
-  background-image: url('../imgs/landingHero.jpg');
+  background-image: url('../assets/imgs/landingHero.jpg');
   background-size: cover;
   background-repeat: no-repeat;
   font-family: 'Montserrat', sans-serif;
